Added dataset utils file
