+++ medinify/datasets/__init__.py
@@ -1,2 +1,3 @@
 from .dataset import Dataset
 from .sentiment_dataset import SentimentDataset
+from .utils import find_csv
