+++ medinify/datasets/dataset.py
@@ -4,6 +4,7 @@ Works with any csv with specified text and label columns
 """
 import pandas as pd
 import os
+from medinify.datasets.utils import find_csv
 
 
 class Dataset:
@@ -113,21 +114,6 @@ class Dataset:
                 str(label), num_label, percent))
         print('\n******************************************************************************************\n')
 
-
-def find_csv(path):
-    """
-    Searches data/csvs directory for data file being loaded
-    :param path: (str) name of csv file being looked for
-    :return absolute_path: (str) (in path found) or None (if not found)
-    """
-    for file in os.walk(os.getcwd()):
-        if os.path.isdir(file[0]) and file[0][-18:] == 'medinify/data/csvs':
-            directory_path = file[0]
-            absolute_path = directory_path + '/' + path
-            if path in os.listdir(directory_path) and os.path.isfile(absolute_path):
-                return absolute_path
-            else:
-                return None
